Fix coordinating inventory across two stock locations
Inventory units assigned to a package cannot also be assigned to another
package or the adjuster will remove them later.

For example there's an order which requires 5 quantities of the same item.
A first stock location has 2 available stock items and
another stock location has 3 available stock items.
Coordinator was creating 5 inventory units and 2 packages, the first
with the first 2 inventory units, the second with the first 3 inventory
units (2 of 3 was shared by both packages).
Prioritizer, when adjusting packages was finding the same inventory
units on both packages removing them from the second one.
The wrong result was a package with 2 items and a package with 1 item.

The fix just stores the inventory units already assigned to a package in
order to build the next packages using only remaining inventory units.