Move checkout coupon code section into summary
Passing params[:order][:coupon_code] is now deprecated so we should not have
this behavior in default frontend code.

This commit moves the coupon code application section of checkout into the
summary box, still visible in payment step only.

It also wraps the coupon code input + button into a real form so it still
works if, for some reason, JS is not working, falling back to deprecated
behavior for now.