Fix address validation switching to a country without states
Saving an address with a country without states, after saving an
address that instead has states will raise a validation error, since
we are not resetting the state. We always had this issue, it's just
raising after #2371, which describe the same issue but for backend.

This commit fixes this problem by adding hidden fields that reset
state_id and/or state_name if they are disabled (via js).