Move sidebar toggle, let top logo always visible
In the first prototype about this new feature, I want to let the Solidus mark
only when the sidebar was collapsed; but then I rembember that the
logo on header is customizable and many times there are no mark (or no
pretty ways) to show in a small place.
So I opted to leave the logo container always visible, also when the user
choose to collapse the sidebar: so, no matter if you've uploaded a custom logo
or not, and furthermore you know always which admin area you are logged in.

The css transition of the collapsing has been improved and applied also to the
toggle button.

Now the language select shows in a smart way when the sidebar is collapsed.