Conditionally raise/validate line item currency mismatches
The current behavior is to raise an exception when the currency
of the price that we are trying to set mismatches with the order
currency.

However, this is creating issues because it's raising also when the
currency is valid. This happens because we don't have control of the
order used by Rails to set attributes on the LineItem instance and
order is still nil/uset when money_price= is executed.

This commit allows to conditionally avoid the raise, switching the
behavior to a standard Rails validation. This is better because at
validation time, we have the instance with all the attributes set
and we are sure to be able to perform the right check.

The validation behavior will become the default with future commits
and the raise bahavior will be deprecated.