Support the selection of currency by customers
A new `Store#currencies` attribute has been added. It contains a set of
currencies that are supported by a store in addition to
`Store#default_currency`.

The currency in the frontend can be switched if it set within the
`'currency'` key in the session. Otherwise, how a default is picked
doesn't change: `Store#default_currency` is taken or, if not present,
`Spree.config.currency`.

A currency switcher has been added to the frontend. In the case a
product has been added to the cart, it's emptied. There're a couple
of reasons why we're not trying to automagically update amounts in the
cart lines:

- Promotions: We can create promotion rules that only apply to a
  currency. It'd be a surprising behavior if we remove a discount to
  the user.
- Customer safety: It's better to explicitly require a user to click on
  the "Add to cart" button while they are seeing the amount of money that
  will be added. Let's imagine that an administrator makes an error and
  sets a super-high price on a product for a specific currency. If a
  user blindly switch to that currency and they perform the checkout,
  they could have a heart attack just afterward.
- The regular flow should be switching currencies and then adding to the
  cart, anyway.

To have a better UX, when requesting a change in the currency an async
call is made to check whether some items have been added to the cart.
When that's the case, the user is prompted to confirm the action.