POC for services arquitecture
This is a POC to show a proposal for the introduction of a service layer
on solidus-core. As it's just a draft, it's not a finished work and
tests and better docs are missing.

As we talked with @kennyadsl, it'd good to be able to use dry-rb tools
as they can help a lot in the introduction of more advanced patterns.
However, the fear existed that it could leave some people out because
the differences with the standard Rails way. This proposal tries to
reach a compromise by introducing a separation between what we can use
internally and what a solidus user needs to know in order to extend its
application. I.e. it uses some libraries internally but it doesn't
require them to extend the framework.

The intersection point between both scenarios is the `Spree::Result`
struct. The idea is that every service on solidus-core must return an
instance of it. Services need to be configurable, so we're going to
provide a default implementation, but users can switch it by their own
one.

The default implementation should be build using
[dry-monads](https://dry-rb.org/gems/dry-monads/). This POC already
applies it to the service to switch a currency. It's not the best
example as it's very simple, but you can get an idea about how powerful
is the library for the kind of logic required for
services/operations/business-transactions reading at its website
(`Maybe`, `Result` & `Try`).

The result of building the service with dry-monads is transformed to a
`Spree::Result` through a `#to_spree_result` method that has been added
as an extension. For that we're using a [dry-core
extension](https://dry-rb.org/gems/dry-core/0.4/extensions/) to show how
it works and makes it a little bit cleaner, but it really wouldn't make
any functional difference to use regular monkey patching instead.

If a user wants to implement its service, they can do as we do and use
dry-monads, or overpass it and do whatever they want as long as they end
up wrapping the result or failure in a `Spree::Result` object:

```ruby
Spree::Result.new(result: SomeResult)
Spree::Result.new(failure: SomeFailure)
```

Using this boundary object between the service layer and its consumer
(the controller) is, in fact, a good thing besides our intention here. Not
exposing all the power of dry-monads to the controller makes sure that
the heavy logic is implemented in the service while the controller only
has access to what is strictly necessary.

`Spree::Result` is internally a
[dry-struct](https://dry-rb.org/gems/dry-core/0.4/extensions/) struct.
It could also be implemented with Ruby's standard library `Struct` or
`OpenStruct` classes, but it has some benefits and avoids some of the
oddities that those default implementations have:

```ruby
Spree::Result.new(result: 'a') == Spree::Result.new(result: 'a') # => true
Struct.new(:result).new(result: 'a') == Struct.new(:result).new(result: 'a') # => false
OpenStruct.new(result: 'a') == OpenStruct.new(result: 'a') # => true

Spree::Result.new(result: 'a').other_method # => NoMethodError
Struct.new(:result).new(result: 'a').other_method # => NoMethodError
OpenStruct.new(result: 'a').other_method # => nil

Spree::Result.new(other_attribute: 'a') # => Dry::Struct::Error
Struct.new(:result).new(:other_attribute) # => #<struct result=:other_attribute>

Spree::Result.new(result: 'a') in {result: the_result} # Pattern matching...
the_result # => 'a'
Struct.new(:result).new(result: 'a') in {result: the_result}
the_result # => 'a'
OpenStruct.new(result: 'a') in {result: the_result} # NoMatchingPatternError
```

dry-struct carries with it
[dry-types](https://dry-rb.org/gems/dry-types/) dependency. As a future
work we could use it to refactor the current type-based configurations,
or even use
[dry-configurable](https://dry-rb.org/gems/dry-configurable/).

In case we accept something like this, a subsequent POC could try to
implement a neater way to inject configurable services.