POC for the injection of services
Building on the POC about introducing services in Solidus, this commit
proposes a mechanism to deal with the injection of user-created
replacements for the Solidus' defaults. However, it's in no way coupled
to the implementation of the former, beyond the expected hierarchy of
files (so no previously used dependencies are needed here).

This POC uses again a dependency from dry-rb:
[dry-system](https://dry-rb.org/gems/dry-system/). Using it makes things
flexible and neat, but if it's too much of a trade-off another option is
to create the same behavior ad hoc for our configuration.

`Solidus::Container` class is responsible for automatically registering
all the services within `core/app/services/spree/`. However, a user can
switch it through a `dependencies` configuration option. Whatever is
finally registered for a given key, can be injected in the controller
thanks to the `Injector` module.

Going through the example in this PR. We have a file
`core/app/services/spree/switch_currency_service.rb` created in the
previous commit. `Solidus::Container` configuration makes it
automatically available under `:switch_currency_service` key. This means
that if we `include Injector[:switch_currency_service]` from
`CurrenciesController`, a new instance of the `SwitchCurrencyService`
class will be available through a
`CurrenciesController#switch_currency_service` method.

If you want to try to substitute the service for something else, you can
try it in the sandbox application. In
`sandbox/config/initializers/spree.rb`, add:

```ruby
Spree.config do |config|
  ....
  config.dependencies do |dependencies|
    dependencies.register(
      :switch_currency_controller, MySwitchCurrencyService.new
    )
  end
  ....
end
```

Then you can create a
`sandbox/app/services/my_switch_currency_service.rb` file with something
like:

```ruby
class MySwitchCurrencyService
  def call(_order)
    abort 'Hello buddy'
  end
end
```

Restart the server and try to switch the currency **having added
something to the cart** to force the new service to be executed.