Add debugging metadata to event objects
We add two attributes to the instances of `Spree::Event::Event` to make
debugging and logging more accessible and useful:

- `time`: Contains the event creation time.
- `caller_location`: Contains an instance of
  `Thread::Backtrace::Location` pointing to the caller of
  `Spree::Event.fire`.

Probably, it'd be a better design to have `Spree::Event.fire` to
provide its caller location instead of defaulting to 4 levels up the
stack on `Spree::Event::Event`. However, as we're still dealing with the
legacy adapter, and it doesn't use `Spree::Event::Event` objects, for
now, it's better to leave things more straightforward to understand.