Fix Spree::Api:LineItemsController#create handling validation errors
Before, #create was expecting that the LineItem instance would have
errors on it if it failed to be created due to validation errors.
However, #add actually raises ActiveRecord::RecordInvalid: Validation
instead, so this endpoint could have never handled validation errors correctly.

This wasn't much of a problem, because LineItem didn't have any
validations that were easy to trigger - the quantity was normalized to
0 if negative, and 0 was allowed, but now 0 is no longer allowed.